+++ standingsrequests/managers/__init__.py
@@ -1,1 +1,0 @@
-REQUIRED_TOKENS = "esi-alliances.read_contacts.v1"